==> 1---Processing_raw_data_2.3.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

df = pd.read_csv('/kaggle/input/match-data/charting-m-points-2020s.csv', encoding='latin-1')
# run this with df = 'charting-m-points-2020.csv' as the data source

dummy_df = df.head(10000)

import pandas as pd

def parse_shots(df):
    # Define character sets for different shot attributes
    shot_types = set('fbrsvzopuylmhikqtS')
    directions = set('123')
    error_types = set('nwdxg!e')
    outcome_types = set('*#@')
    depth_types = set('789')
    serve_zones = set('456')

    def parse_code(code):
        shots = []
        shot = {'shot_num': 1, 'shot_type': '', 'direction': '', 'error_type': '', 'outcome': '', 'depth': '', 'position': '', 'net_cord': '', 'approach': '', 'serve_and_volley': False, 'serve_zone': ''}
        i = 0
        
        while i < len(code):
            char = code[i]
            
            if char in serve_zones:
                shot = {'shot_num': 1, 'shot_type': 'S', 'direction': '', 'error_type': '', 'outcome': '', 'depth': '', 'position': '', 'net_cord': '', 'approach': '', 'serve_and_volley': False, 'serve_zone': char}
            
            if char in shot_types:
                shots.append(shot)
                shot = {'shot_num': len(shots) + 1, 'shot_type': char, 'direction': '', 'error_type': '', 'outcome': '', 'depth': '', 'position': '', 'net_cord': '', 'approach': '', 'serve_and_volley': False, 'serve_zone': ''}
                
            elif char in directions:
                shot['direction'] = char

            elif char in error_types:
                shot['error_type'] = char

            elif char in outcome_types:
                shot['outcome'] = char

            elif char in depth_types:
                shot['depth'] = char

            elif char == '+':
                shot['approach'] = '+'

            elif char == '-':
                shot['position'] = '-'

            elif char == '=':
                shot['position'] = '='

            elif char == ';':
                shot['net_cord'] = ';'

            elif char == 'c':
                shot['serve_and_volley'] = True

            i += 1
        
        if shot['shot_type']:
            shots.append(shot)

        return shots

    data = {col: [] for col in df.columns}
    extra_columns = ['serve_zone', 'shot_type', 'direction', 'error_type', 'outcome', 'depth', 'position', 'net_cord', 'approach', 'serve_and_volley', 'shot_num']
    for col in extra_columns:
        data[col] = []

    for index, row in df.iterrows():
        for col in ['1st', '2nd']:
            if pd.notna(row[col]):
                shots = parse_code(row[col])
                for shot in shots:
                    for df_col in df.columns:
                        data[df_col].append(row[df_col])
                    data['serve_zone'].append(shot['serve_zone'])
                    data['shot_type'].append(shot['shot_type'])
                    data['direction'].append(shot['direction'])
                    data['error_type'].append(shot['error_type'])
                    data['outcome'].append(shot['outcome'])
                    data['depth'].append(shot['depth'])
                    data['position'].append(shot['position'])
                    data['net_cord'].append(shot['net_cord'])
                    data['approach'].append(shot['approach'])
                    data['serve_and_volley'].append(shot['serve_and_volley'])
                    data['shot_num'].append(shot['shot_num'])

    return pd.DataFrame(data)

def add_player_shot_column(df):
    match_components = df['match_id'].str.split('-', expand=True)
    df['Player_1'] = match_components.iloc[:, -2]  # Second to last component
    df['Player_2'] = match_components.iloc[:, -1]  # Last component
    
    df['Player_1'] = df['Player_1'].str.replace('_', ' ')
    df['Player_2'] = df['Player_2'].str.replace('_', ' ')

    df['Hitting_Player'] = None
    
    current_server = None
    current_shot_num = None

    for index, row in df.iterrows():
        if row['shot_num'] == 1:
            current_server = row['Svr']
            current_shot_num = 1
        
        if current_shot_num % 2 == 1:
            hitting_player = df.at[index, 'Player_1'] if current_server == 1 else df.at[index, 'Player_2']
        else:
            hitting_player = df.at[index, 'Player_2'] if current_server == 1 else df.at[index, 'Player_1']
        
        df.at[index, 'Hitting_Player'] = hitting_player
        
        current_shot_num += 1
    
    return df

def add_identifiers(df):
    # Initialize the columns
    df['pt_id'] = None
    df['gm_id'] = None
    df['set_id'] = None

    # Variables to keep track of current set and game
    current_set = None
    current_match = None

    # Initialize dictionaries to track set and game identifiers
    set_ids = {}
    game_ids = {}

    for idx, row in df.iterrows():
        match_id = row['match_id']
        pt_number = row['Pt']
        game_number = row['Gm#'][0]  # Assuming Gm# is a list, make sure this is correct
        
        # Update current match if changed
        if match_id != current_match:
            current_match = match_id
            # Reset current set for a new match
            current_set = None
        
        # Check for changes in sets
        new_set = row['Set1'] + row['Set2'] + 1  # Alternatively, check the actual set being played
        if new_set != current_set:
            current_set = new_set
            set_ids[current_set] = f"{match_id}_Set{current_set}"
        
        # Check for changes in games
        if (current_set, game_number) not in game_ids:
            game_ids[(current_set, game_number)] = f"{match_id}_Set{current_set}_Game{game_number}"
        
        # Assign identifiers
        pt_id = f"{match_id}_Pt{pt_number}"
        gm_id = game_ids.get((current_set, game_number), None)
        set_id = set_ids.get(current_set, None)
        
        df.at[idx, 'pt_id'] = pt_id
        df.at[idx, 'gm_id'] = gm_id
        df.at[idx, 'set_id'] = set_id

    return df

def determine_winners(df):
    # Ensure PtWinner_Player column is available
    df['PtWinner_Player'] = df.apply(lambda row: row['Player_1'] if row['PtWinner'] == 1 else row['Player_2'], axis=1)
    
    # Initialize columns for winners
    df['GmWinner_Player'] = None
    df['SetWinner_Player'] = None
    df['MatchWinner_Player'] = None
    
    # Process match by match
    for match_id, match_df in df.groupby('match_id'):
        # Determine Match Winner from the last point of the match
        match_winner = match_df['PtWinner_Player'].iloc[-1]
        df.loc[match_df.index, 'MatchWinner_Player'] = match_winner
        
        # Process set by set within the match
        for set_id, set_df in match_df.groupby('set_id'):
            # Determine Set Winner from the last point of the set
            set_winner = set_df['PtWinner_Player'].iloc[-1]
            df.loc[set_df.index, 'SetWinner_Player'] = set_winner
            
            # Process game by game within the set
            for gm_id, gm_df in set_df.groupby('gm_id'):
                # Determine Game Winner from the last point of the game
                game_winner = gm_df['PtWinner_Player'].iloc[-1]
                df.loc[gm_df.index, 'GmWinner_Player'] = game_winner

    return df

def add_shot_in_column(df):
    # Ensure that PtWinner_Player column is available
    df['PtWinner_Player'] = df.apply(lambda row: row['Player_1'] if row['PtWinner'] == 1 else row['Player_2'], axis=1)
    
    # Initialize the 'shot_in' column as 1 (meaning all shots are in by default)
    df['shot_in'] = 1

    # Group by both match_id and Pt to ensure the logic applies to each point within each match
    for (match_id, pt), group in df.groupby(['match_id', 'Pt']):
        first_shot_index = group.index[0]
        second_shot_index = first_shot_index + 1 if len(group) > 1 else None
        
        # Check if the first serve is in or out
        if group.at[first_shot_index, '1stIn'] == 0:
            # First serve is out
            df.at[first_shot_index, 'shot_in'] = 0
            
            # Check if second serve exists and whether it's in or out
            if second_shot_index in group.index and group.at[second_shot_index, 'shot_num'] == 1:
                if group.at[second_shot_index, '2ndIn'] == 0:
                    # Second serve is out, double fault
                    df.at[second_shot_index, 'shot_in'] = 0
                else:
                    # Second serve is in
                    df.at[second_shot_index, 'shot_in'] = 1

        # If the first serve is in
        else:
            df.at[first_shot_index, 'shot_in'] = 1

        # Handle the rest of the rally after the serve(s)
        last_shot_index = group.index[-1]
        last_hitting_player = group.loc[last_shot_index, 'Hitting_Player']
        pt_winner_player = group.loc[last_shot_index, 'PtWinner_Player']

        # Mark the last shot of the rally as out if the loser of the point hit it
        if last_hitting_player != pt_winner_player:
            df.at[last_shot_index, 'shot_in'] = 0
    
    return df


def split_pts_column(df):
    
    def assign_points(row):
        scores = row['Pts'].split('-')
        
        if len(scores) != 2:
            return pd.Series([None, None], index=['Player_1_Pts', 'Player_2_Pts'])
        
        player_1_pts, player_2_pts = str(scores[0]), str(scores[1])
        
        if row['Svr'] == 2:
            player_1_pts, player_2_pts = player_2_pts, player_1_pts
            
        return pd.Series([player_1_pts, player_2_pts], index=['Player_1_Pts', 'Player_2_Pts'])
    
    df[['Player_1_Pts', 'Player_2_Pts']] = df.apply(assign_points, axis=1)

    # Move Player_1_Pts and Player_2_Pts to indices 7 and 8
    cols = df.columns.tolist()  # Get the current column order
    # Remove the columns from their current positions
    cols.remove('Player_1_Pts')
    cols.remove('Player_2_Pts')
    
    # Insert Player_1_Pts and Player_2_Pts at indices 7 and 8
    cols.insert(7, 'Player_1_Pts')
    cols.insert(8, 'Player_2_Pts')
    
    # Reorder the DataFrame
    df = df[cols]

    return df

parsed_df = parse_shots(df)
parsed_df = add_player_shot_column(parsed_df)
parsed_df = add_identifiers(parsed_df)
parsed_df = determine_winners(parsed_df)
parsed_df = add_shot_in_column(parsed_df)
final_df = split_pts_column(parsed_df)
final_df.to_csv('preprocessed_aggressiveness.csv', index=False)

final_df
